Add provider and split some pages into submodules
