+++ lib/home.dart
@@ -1,0 +1,71 @@
+import 'package:flutter/material.dart';
+import 'pages.dart';
+
+class Home extends StatefulWidget {
+  const Home({super.key});
+
+  @override
+  HomeState createState() => HomeState();
+}
+
+class HomeState extends State<Home> {
+  int screenIndex = 0;
+
+  void handleScreenChanged(int index) {
+    setState(() {
+      screenIndex = index;
+    });
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(
+        leading: const Icon(Icons.menu),
+        title: const Text('LatinOne'),
+        backgroundColor: Colors.orange,
+        centerTitle: true,
+        actions: const <Widget>[
+          IconButton(
+            icon: Icon(Icons.email, color: Colors.white),
+            onPressed: null, //TODO:実装
+          ),
+          IconButton(
+            icon: Icon(Icons.settings, color: Colors.white),
+            onPressed: null, //TODO:実装
+          ),
+        ],
+      ),
+
+      //Body
+      body: SelectedScreen.values[screenIndex].create(),
+
+      // BottomNavigationBar
+      bottomNavigationBar: BottomNavigationBar(
+        onTap: (index) {
+          setState(() {
+            screenIndex = index;
+            handleScreenChanged(screenIndex);
+          });
+        },
+        currentIndex: screenIndex,
+        items: const <BottomNavigationBarItem>[
+          BottomNavigationBarItem(
+            icon: Icon(Icons.home),
+            label: 'Home',
+          ),
+          BottomNavigationBarItem(
+            icon: Icon(Icons.store),
+            label: 'Store',
+          ),
+          BottomNavigationBarItem(
+            icon: Icon(Icons.local_shipping),
+            label: 'Order',
+          ),
+        ],
+        fixedColor: Colors.blueAccent,
+        type: BottomNavigationBarType.fixed,
+      ),
+    );
+  }
+}
