+++ lib/product.dart
@@ -1,0 +1,15 @@
+class Product {
+  final int    id;
+  final String name;
+  final String description;
+  final int    price;
+  final String image;
+
+  const Product(
+    this.id,
+    this.name,
+    this.description,
+    this.price,
+    this.image
+  );
+}
