+++ pubspec.yaml
@@ -37,6 +37,7 @@ dependencies:
   cupertino_icons: ^1.0.6
   flutter_map: ^7.0.2
   latlong2: ^0.9.1
+  provider: ^6.1.2
 
 dev_dependencies:
   flutter_test:
